Implement signal strength calculation with double precision

+++ src/Aeromux.Core/ModeS/Messages/AllCallReply.cs
@@ -30,7 +30,7 @@ namespace Aeromux.Core.ModeS.Messages;
 /// <param name="IcaoAddress">ICAO aircraft address.</param>
 /// <param name="Timestamp">UTC timestamp when the message was received.</param>
 /// <param name="DownlinkFormat">Downlink format (DF 11).</param>
-/// <param name="SignalStrength">Signal strength in dBFS (0-255).</param>
+/// <param name="SignalStrength">Signal strength in RSSI (0-255).</param>
 /// <param name="WasCorrected">True if error correction was applied.</param>
 /// <param name="ExtractedIcao">Extracted ICAO from the AA field Should be same as the IcaoAddress.</param>
 /// <param name="Capability">Transponder capability level (Mode S support level and status).</param>
@@ -38,7 +38,7 @@ public sealed record AllCallReply(
     string IcaoAddress,
     DateTime Timestamp,
     DownlinkFormat DownlinkFormat,
-    byte SignalStrength,
+    double SignalStrength,
     bool WasCorrected,
     string ExtractedIcao,
     TransponderCapability Capability) : ModeSMessage(IcaoAddress, Timestamp, DownlinkFormat, SignalStrength, WasCorrected);
